🛠 add dry run mode
See #12

+++ src/google-calendar-client.js
@@ -1,5 +1,11 @@
 import { request, parseBodyAsJson } from "./http-client.js";
 
+const DRY_RUN = process.env.DRY_RUN === "true";
+
+if (DRY_RUN) {
+  console.warn("WARN: DRY_RUN is set to 'true'! Nothing will be written to Google Calendar.")
+}
+
 export async function getCalendar(calendarId, accessToken) {
   const response = await request({
     url: `https://www.googleapis.com/calendar/v3/calendars/${calendarId}`,
@@ -13,6 +19,9 @@ export async function getCalendar(calendarId, accessToken) {
 }
 
 export async function insertEvent(calendarId, eventBody, accessToken) {
+  if (DRY_RUN) {
+    return;
+  }
   const response = await request({
     url: `https://www.googleapis.com/calendar/v3/calendars/${calendarId}/events`,
     method: "POST",
@@ -27,6 +36,9 @@ export async function insertEvent(calendarId, eventBody, accessToken) {
 }
 
 export async function updateEvent(calendarId, eventId, eventBody, accessToken) {
+  if (DRY_RUN) {
+    return;
+  }
   const response = await request({
     url: `https://www.googleapis.com/calendar/v3/calendars/${calendarId}/events/${eventId}`,
     method: "PUT",
@@ -41,6 +53,9 @@ export async function updateEvent(calendarId, eventId, eventBody, accessToken) {
 }
 
 export async function removeEvent(calendarId, eventId, accessToken) {
+  if (DRY_RUN) {
+    return;
+  }
   const response = await request({
     url: `https://www.googleapis.com/calendar/v3/calendars/${calendarId}/events/${eventId}`,
     method: "DELETE",
